Rewards: refactor to have common GymHardcodedReward base class
